refactor: decompose monolithic _op.py into focused modules
NGRAF-006: Split 1378-line _op.py into 4 specialized modules while
maintaining full backward compatibility.

Changes:
- Create chunking.py (115 lines) for text chunking operations
- Create extraction.py (466 lines) for entity/relationship extraction
- Create community.py (383 lines) for community detection and reports
- Create query.py (435 lines) for local/global/naive query operations
- Transform _op.py into 55-line compatibility layer with deprecation warning
- Update graphrag.py to import from new modules directly
- Add comprehensive unit tests for each new module (29 tests total)

Benefits:
- Improved code organization and single responsibility principle
- Enhanced maintainability with smaller, focused modules
- Better testability with module-specific test files
- Zero breaking changes - all existing code continues to work
- Clear migration path with deprecation warnings

All 141 existing tests pass without modification. The refactoring
preserves all global_config patterns and function signatures to
ensure compatibility. Phase 2 improvements (removing global_config
dependencies) will be addressed in a future PR.
