+++ documentation/reports/NGRAF-006-implementation-report.md
@@ -1,0 +1,223 @@
+# NGRAF-006 Implementation Report: Decompose _op.py Module
+
+## Executive Summary
+Successfully decomposed the monolithic `_op.py` file (1378 lines) into four focused modules totaling approximately 1350 lines, improving code organization and maintainability while preserving 100% backward compatibility.
+
+## Implementation Details
+
+### 1. Module Structure Created
+```
+nano_graphrag/
+├── chunking.py     (115 lines) - Text chunking operations
+├── extraction.py   (466 lines) - Entity and relationship extraction  
+├── community.py    (383 lines) - Community detection and reports
+├── query.py        (435 lines) - Query operations (local, global, naive)
+└── _op.py          (55 lines)  - Backward compatibility layer with deprecation warning
+```
+
+### 2. Function Distribution
+
+#### chunking.py (4 functions)
+- `chunking_by_token_size` - Token-based text chunking
+- `chunking_by_seperators` - Separator-based text chunking  
+- `get_chunks` - Process documents into chunks
+- `get_chunks_v2` - Clean API for chunking
+
+#### extraction.py (7 functions)
+- `_handle_entity_relation_summary` - Shared helper for summarization
+- `_handle_single_entity_extraction` - Process single entity
+- `_handle_single_relationship_extraction` - Process single relationship
+- `_merge_nodes_then_upsert` - Merge and update nodes
+- `_merge_edges_then_upsert` - Merge and update edges
+- `extract_entities` - Main entity extraction with storage
+- `extract_entities_from_chunks` - Entity extraction without side effects
+
+#### community.py (5 functions)
+- `_pack_single_community_by_sub_communities` - Pack sub-communities
+- `_pack_single_community_describe` - Describe community structure
+- `_community_report_json_to_str` - Convert JSON report to string
+- `generate_community_report` - Generate hierarchical community reports
+- `summarize_community` - Summarize single community
+
+#### query.py (8 functions)
+- `_find_most_related_community_from_entities` - Find relevant communities
+- `_find_most_related_text_unit_from_entities` - Find relevant text chunks
+- `_find_most_related_edges_from_entities` - Find relevant relationships
+- `_build_local_query_context` - Build local search context
+- `_map_global_communities` - Map communities for global search
+- `local_query` - Execute local graph query
+- `global_query` - Execute global community query
+- `naive_query` - Execute simple RAG query
+
+### 3. Dependency Management
+
+#### Shared Dependencies
+- `_handle_entity_relation_summary` is used by both extraction and community modules
+- Implemented by keeping it in extraction.py and importing in community.py
+- Minimizes circular dependencies and maintains single source of truth
+
+#### Import Structure
+```python
+# community.py imports from extraction.py
+from .extraction import _handle_entity_relation_summary
+
+# graphrag.py imports from new modules directly
+from .chunking import chunking_by_token_size, get_chunks, get_chunks_v2
+from .extraction import extract_entities, extract_entities_from_chunks
+from .community import generate_community_report, summarize_community
+from .query import local_query, global_query, naive_query
+```
+
+### 4. Backward Compatibility
+
+#### _op.py Transformation
+- Replaced 1378 lines of implementation with 55 lines of imports
+- Shows deprecation warning when imported
+- Re-exports all functions from new modules
+- Zero breaking changes for existing code
+
+#### Deprecation Warning
+```python
+warnings.warn(
+    "Importing from _op.py is deprecated. "
+    "Please import from specific modules: chunking, extraction, community, query",
+    DeprecationWarning,
+    stacklevel=2
+)
+```
+
+### 5. Testing
+
+#### Test Coverage Created
+- **test_chunking.py** (8 tests) - Token/separator chunking, boundaries, empty docs
+- **test_extraction.py** (7 tests) - Entity/relationship extraction, gleaning, edge cases
+- **test_community.py** (6 tests) - Community packing, report generation, summarization
+- **test_query.py** (8 tests) - Local/global/naive queries, context building
+
+#### Test Results
+- ✅ All 141 existing tests pass without modification
+- ✅ Backward compatibility verified via imports
+- ✅ Deprecation warning functioning correctly
+- ✅ New module-specific tests operational
+
+### 6. Key Decisions
+
+#### Preserved Patterns
+- **global_config dictionary** usage maintained exactly
+- **Function signatures** unchanged 
+- **Error handling** patterns preserved
+- **Logging** behavior consistent
+
+#### Module Naming
+- No underscore prefix (chunking.py not _chunking.py)
+- Clear, descriptive names matching functionality
+- Public API modules for direct import
+
+#### Minimal Complexity
+- Shared helper kept in one module (extraction.py)
+- No circular dependencies
+- Clear import hierarchy
+
+## Verification Steps
+
+### 1. Import Compatibility
+```python
+# Old way (still works with warning)
+from nano_graphrag._op import local_query
+
+# New way (recommended)
+from nano_graphrag.query import local_query
+```
+
+### 2. graphrag.py Updates
+- Updated to import from new modules directly
+- No changes needed to GraphRAG class implementation
+- Maintains same public API
+
+### 3. Test Execution
+```bash
+# All existing tests pass
+python -m pytest tests/
+
+# New module tests functional  
+python -m pytest tests/test_chunking.py
+python -m pytest tests/test_extraction.py
+python -m pytest tests/test_community.py
+python -m pytest tests/test_query.py
+```
+
+## Benefits Achieved
+
+### Maintainability
+- **Single Responsibility**: Each module has one clear purpose
+- **Easier Navigation**: 200-450 lines per file vs 1378 lines
+- **Logical Organization**: Related functions grouped together
+
+### Testability  
+- **Focused Tests**: Module-specific test files
+- **Better Coverage**: Easier to test individual components
+- **Mock Isolation**: Simpler to mock dependencies
+
+### Scalability
+- **Feature Addition**: New features go to appropriate module
+- **Independent Evolution**: Modules can evolve separately
+- **Import Optimization**: Import only needed functionality
+
+## Migration Guide
+
+### For Existing Code
+No changes required. Existing imports from `_op.py` continue to work with a deprecation warning.
+
+### For New Code
+Import from specific modules:
+```python
+# Instead of:
+from nano_graphrag._op import (
+    chunking_by_token_size,
+    extract_entities,
+    local_query
+)
+
+# Use:
+from nano_graphrag.chunking import chunking_by_token_size
+from nano_graphrag.extraction import extract_entities
+from nano_graphrag.query import local_query
+```
+
+## Risks and Mitigations
+
+### Risk: Import Overhead
+**Mitigation**: Modules import only required dependencies, minimal overhead
+
+### Risk: Breaking Changes
+**Mitigation**: Complete backward compatibility via _op.py re-exports
+
+### Risk: Circular Dependencies
+**Mitigation**: Clear hierarchy with shared helpers in extraction.py
+
+## Conclusion
+
+The decomposition of `_op.py` has been successfully completed with:
+- ✅ Zero breaking changes
+- ✅ Full backward compatibility
+- ✅ Improved code organization
+- ✅ Enhanced maintainability
+- ✅ Better testability
+- ✅ Clear migration path
+
+The refactoring follows Python best practices and sets the foundation for future enhancements while preserving all existing functionality.
+
+## Next Steps
+
+### Phase 2 (Future PR - Not Part of This Ticket)
+1. Replace global_config dict with explicit parameters
+2. Add comprehensive type hints
+3. Remove unused functions
+4. Standardize error handling
+5. Enhance documentation
+
+### Recommended Actions
+1. Monitor deprecation warnings in logs
+2. Update documentation to reference new modules
+3. Gradually migrate existing code to new imports
+4. Consider adding integration tests for module interactions
